Move workflow helpers to "utils" folder
