+++ src/ggrc-client/js/components/workflow/cycle-task-group-object-task/tests/cycle-task-group-object-task_spec.js
@@ -5,7 +5,7 @@
 
 import Component from '../cycle-task-group-object-task';
 import RefreshQueue from '../../../../models/refresh_queue';
-import WorkflowHelpers from '../../workflow-helpers';
+import WorkflowHelpers from '../../../../plugins/utils/workflow-utils';
 
 describe('GGRC.Components.cycleTaskGroupObjectTask', function () {
   let viewModel;
