+++ src/ggrc-client/js/components/workflow/workflow-activate.js
@@ -3,7 +3,7 @@
   Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import workflowHelpers from './workflow-helpers';
+import workflowHelpers from '../../plugins/utils/workflow-utils';
 import {
   initCounts,
 } from '../../plugins/utils/current-page-utils';
