+++ src/ggrc-client/js/components/workflow/workflow-start-cycle.js
@@ -3,7 +3,7 @@
   Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import workflowHelpers from './workflow-helpers';
+import workflowHelpers from '../../plugins/utils/workflow-utils';
 
 export default can.Component.extend({
   tag: 'workflow-start-cycle',
