+++ src/ggrc-client/js/plugins/tests/utils/workflow-utils_spec.js
@@ -3,7 +3,7 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import workflowHelpers from '../workflow-helpers';
+import workflowHelpers from '../../utils/workflow-utils';
 import {
   makeFakeInstance,
   makeFakeModel,
