+++ src/ggrc-client/js/plugins/utils/workflow-utils.js
